move coremod, change mixin package name
