+++ src/main/java/ruiseki/okbackpack/mixins/early/minecraft/MixinGuiContainerBackpackInteraction.java
@@ -1,4 +1,4 @@
-package ruiseki.okbackpack.mixins.early.Minecraft;
+package ruiseki.okbackpack.mixins.early.minecraft;
 
 import java.util.List;
