+++ src/main/java/ruiseki/okbackpack/mixins/early/minecraft/MixinGuiContainerBackpackOpener.java
@@ -1,4 +1,4 @@
-package ruiseki.okbackpack.mixins.early.Minecraft;
+package ruiseki.okbackpack.mixins.early.minecraft;
 
 import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.client.gui.inventory.GuiContainer;
