+++ src/main/java/ruiseki/okbackpack/mixins/early/minecraft/MixinPlayerControllerMPBackpackInteraction.java
@@ -1,4 +1,4 @@
-package ruiseki.okbackpack.mixins.early.Minecraft;
+package ruiseki.okbackpack.mixins.early.minecraft;
 
 import net.minecraft.client.multiplayer.PlayerControllerMP;
 import net.minecraft.client.network.NetHandlerPlayClient;
