+++ src/main/java/ruiseki/okbackpack/mixins/late/thaumcraft/MixinTileMagicWorkbench.java
@@ -1,4 +1,4 @@
-package ruiseki.okbackpack.mixins.late.Thaumcraft;
+package ruiseki.okbackpack.mixins.late.thaumcraft;
 
 import java.util.Collections;
